requirements.txt
# Django and REST Framework
Django==5.1.7
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==24.1
django-celery-results==2.5.1
django-environ==0.11.2
django-debug-toolbar==5.1.0 


# Database
psycopg2-binary==2.9.9

# Async tasks
celery==5.3.6
redis==5.0.1

# ML and data processing
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
transformers>=4.40.0
sentence-transformers>=3.0.0
networkx==3.3.0

# File handling
openpyxl==3.1.2
pdfplumber>=0.5.28
python-docx==1.1.2

# Testing
pytest==8.0.0
pytest-django==4.7.0

# Production
gunicorn==21.2.0

# Development tools
black==24.1.1
isort==5.13.2
flake8==7.0.0
